# documentazione/swagger-openapi.yaml
openapi: 3.0.1
info:
  title: OSP API
  description: Gestionale Occupazione Suolo Pubblico Bari
  version: '1.0'
servers:
  - url: http://localhost:8080/api/osp
    description: Generated server url
paths:
...
  /tipologiche/tipologie-titoli-edilizi:
    get:
      tags:
        - typological-controller
      operationId: getTipologieTitoliEdilizi
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TypologicalFlagTestoLiberoDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /tipologiche/tipi-template:
    get:
      tags:
        - typological-controller
      operationId: getTipiTemplate
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TypologicalDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /tipologiche/tipi-ruoli-richiedenti:
    get:
      tags:
        - typological-controller
      operationId: getTipiRuoliRichiedenti
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TypologicalDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /tipologiche/tipi-processi:
    get:
      tags:
        - typological-controller
      operationId: getTipiProcessi
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TypologicalDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /tipologiche/tipi-notifiche-scadenzario:
    get:
      tags:
        - typological-controller
      operationId: getTipiNotificheScadenzario
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TypologicalDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /tipologiche/tipi-manufatti:
    get:
      tags:
        - typological-controller
      operationId: getTipiManufatti
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TypologicalFlagTestoLiberoDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /tipologiche/tipi-documenti-allegati:
    get:
      tags:
        - typological-controller
      operationId: getTipiDocumentiAllegati
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TypologicalDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /tipologiche/tipi-attivita-da-svolgere:
    get:
      tags:
        - typological-controller
      operationId: getTipiAttivitaDaSvolgere
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TypologicalFlagTestoLiberoDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /tipologiche/tipi-allegati:
    get:
      tags:
        - typological-controller
      operationId: getTipiAllegati
      parameters:
        - name: idUtente
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: idPratica
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TipoAllegatoGruppoStatoProcessoDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /tipologiche/tipi-allegati/all:
    get:
      tags:
        - typological-controller
      operationId: getAllTipiAllegati
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TipoAllegatoDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /tipologiche/tipi-allegati-stato-processo:
    get:
      tags:
        - typological-controller
      operationId: getTipiAllegatiByStatoAndProcesso
      parameters:
        - name: idUtente
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: idStatoPratica
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: idTipoProcesso
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TipoAllegatoGruppoStatoProcessoDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /tipologiche/stati-pratiche:
    get:
      tags:
        - typological-controller
      operationId: getStatiPratiche
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TypologicalDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /tipologiche/municipi:
    get:
      tags:
        - typological-controller
      operationId: getMunicipi
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TypologicalDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /tipologiche/gruppi:
    get:
      tags:
        - typological-controller
      operationId: getGruppi
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GruppoDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
...
components:
  schemas:
...
    TypologicalDTO:
      required:
        - descrizione
        - descrizioneEstesa
        - id
      type: object
      properties:
        id:
          type: integer
          format: int32
        descrizione:
          type: string
        descrizioneEstesa:
          type: string
...
    TipoAllegatoDTO:
      required:
        - descrizione
        - id
      type: object
      properties:
        id:
          type: integer
          format: int32
        descrizione:
          type: string
        descrizioneEstesa:
          type: string
...
    TypologicalFlagTestoLiberoDTO:
      required:
        - descrizione
        - descrizioneEstesa
        - id
      type: object
      properties:
        id:
          type: integer
          format: int32
        descrizione:
          type: string
        descrizioneEstesa:
          type: string
        flagTestoLibero:
          type: boolean
    TipoAllegatoGruppoStatoProcessoDTO:
      type: object
      properties:
        tipoAllegato:
          $ref: '#/components/schemas/TipoAllegatoDTO'
        flagObbligatorio:
          type: boolean
        flagTestoLibero:
          type: boolean
        idStatoPratica:
          type: integer
          format: int32
        idTipoProcesso:
          type: integer
          format: int32
        idGruppo:
          type: integer
          format: int32
    GruppoDTO:
      required:
        - descrizione
        - id
      type: object
      properties:
        id:
          type: integer
          format: int32
        descrizione:
          type: string
        flagGestInvioMailDetermina:
          type: boolean
        flagGestRichiestaParere:
          type: boolean
        flagInserisciRichiesta:
          type: boolean
        flagConcessioniValide:
          type: boolean
        flagFascicolo:
          type: boolean
        flagPraticaInserita:
          type: boolean
        flagPresaInCarico:
          type: boolean
        flagVerificaPratiche:
          type: boolean
        flagRichiestaPareri:
          type: boolean
        flagNecessariaIntegrazione:
          type: boolean
        flagPraticheApprovate:
          type: boolean
        flagAttesaPagamento:
          type: boolean
        flagPronteRilascio:
          type: boolean
        flagPraticheDaRigettare:
          type: boolean
        flagPratichePreavvisoDiniego:
          type: boolean
        flagPraticheArchiviate:
          type: boolean
        flagPraticheRigettate:
          type: boolean
        flagGestioneProfilo:
          type: boolean
        flagGestioneUtenti:
          type: boolean
        flagVerificaRipristinoLuoghi:
          type: boolean
...
  securitySchemes:
    jwt:
      type: http
      in: header
      scheme: Bearer